CallTimer: Fix lag
